move Task and initialize_tasks() into their own task module
