+++ phase1-coordinator/src/coordinator_state.rs
@@ -1,6 +1,9 @@
 use crate::{
     environment::Environment,
-    objects::participant::*,
+    objects::{
+        participant::*,
+        task::{initialize_tasks, Task},
+    },
     storage::{Locator, Object, StorageLock},
     CoordinatorError,
     TimeSource,
@@ -9,16 +12,10 @@ use phase1::ProvingSystem;
 
 use chrono::{DateTime, Duration, Utc};
 use rayon::prelude::*;
-use serde::{
-    de::{self, Deserializer, Error},
-    ser::Serializer,
-    Deserialize,
-    Serialize,
-};
+use serde::{Deserialize, Serialize};
 use std::{
     collections::{HashMap, HashSet, LinkedList},
     iter::FromIterator,
-    str::FromStr,
 };
 use tracing::*;
 
@@ -1974,7 +1971,7 @@ impl CoordinatorState {
                     .pending_tasks
                     .iter()
                     .cloned()
-                    .partition(|task| tasks_by_chunk.get(&task.chunk_id).is_some());
+                    .partition(|task| tasks_by_chunk.get(&task.chunk_id()).is_some());
 
                 contributor_info.disposing_tasks = disposing_tasks;
                 contributor_info.pending_tasks = pending_tasks;
@@ -1982,8 +1979,8 @@ impl CoordinatorState {
                 // If completed task is based on the dropped task, it should also be dropped
                 let (disposed_tasks, completed_tasks) =
                     contributor_info.completed_tasks.iter().cloned().partition(|task| {
-                        if let Some(contribution_id) = tasks_by_chunk.get(&task.chunk_id) {
-                            *contribution_id < task.contribution_id
+                        if let Some(contribution_id) = tasks_by_chunk.get(&task.chunk_id()) {
+                            *contribution_id < task.contribution_id()
                         } else {
                             false
                         }
@@ -1995,14 +1992,14 @@ impl CoordinatorState {
 
                 // Determine the excluded tasks, which are filtered out from the list of newly assigned tasks.
                 let mut excluded_tasks: HashSet<u64> =
-                    HashSet::from_iter(contributor_info.completed_tasks.iter().map(|task| task.chunk_id));
-                excluded_tasks.extend(contributor_info.pending_tasks.iter().map(|task| task.chunk_id));
+                    HashSet::from_iter(contributor_info.completed_tasks.iter().map(|task| task.chunk_id()));
+                excluded_tasks.extend(contributor_info.pending_tasks.iter().map(|task| task.chunk_id()));
 
                 // Reassign tasks for the affected contributor.
                 contributor_info.assigned_tasks =
                     initialize_tasks(contributor_info.bucket_id, number_of_chunks, number_of_contributors)
                         .into_iter()
-                        .filter(|task| !excluded_tasks.contains(&task.chunk_id))
+                        .filter(|task| !excluded_tasks.contains(&task.chunk_id()))
                         .collect();
             }
 
@@ -2049,7 +2046,7 @@ impl CoordinatorState {
             // Add just the current pending tasks to a pending verifications list.
             let mut pending_verifications = vec![];
             for task in &tasks {
-                pending_verifications.push((task.chunk_id, task.contribution_id));
+                pending_verifications.push((task.chunk_id(), task.contribution_id()));
             }
 
             // Set the participant as dropped.
@@ -3051,61 +3048,6 @@ impl CoordinatorState {
     }
 }
 
-fn initialize_tasks(
-    starting_bucket_id: u64, // 0-indexed
-    number_of_chunks: u64,
-    number_of_contributors: u64,
-) -> LinkedList<Task> {
-    let number_of_buckets = number_of_contributors;
-
-    // Fetch the bucket size.
-    let bucket_size = number_of_chunks / number_of_buckets as u64;
-
-    // It takes `number_of_contributors * bucket_size` to get to initial stable state.
-    // You will jump up (total_jumps - starting_bucket_id) times.
-    let number_of_initial_steps = (number_of_contributors - 1) - starting_bucket_id;
-    let number_of_final_steps = starting_bucket_id;
-    let mut initial_steps = 0;
-    let mut final_steps = 0;
-
-    // Compute the start and end indices.
-    let start = starting_bucket_id * bucket_size;
-    let end = start + number_of_chunks;
-
-    // Add the tasks in FIFO ordering.
-    let mut tasks = LinkedList::new();
-    tasks.push_back(Task::new(start % number_of_chunks, 1));
-
-    // Skip one from the start index and calculate the chunk ID and contribution ID to the end.
-    for current_index in start + 1..end {
-        let chunk_id = current_index % number_of_chunks;
-
-        // Check if this is a new bucket.
-        let is_new_bucket = (chunk_id % bucket_size) == 0;
-
-        // Check if we have initial steps to increment.
-        if is_new_bucket && initial_steps < number_of_initial_steps {
-            initial_steps += 1;
-        }
-
-        // Check if we have iterated past the modulus and are in final steps.
-        if is_new_bucket && current_index >= number_of_chunks {
-            // Check if we have final steps to increment.
-            if final_steps < number_of_final_steps {
-                final_steps += 1;
-            }
-        }
-
-        // Compute the contribution ID.
-        let steps = (initial_steps + final_steps) % number_of_contributors;
-        let contribution_id = steps + 1;
-
-        tasks.push_back(Task::new(chunk_id, contribution_id as u64));
-    }
-
-    tasks
-}
-
 /// Data required by the coordinator to drop a participant from the
 /// ceremony.
 pub(crate) struct DropData {
@@ -3135,131 +3077,9 @@ pub(crate) enum Justification {
     Inactive,
 }
 
-/// The id of a task to be performed by a ceremony participant at a
-/// given contribution level, for a given chunk.
-///
-/// ```txt, ignore
-/// +----------------+---------+---------+---------+
-/// | ...            |  Task   |  Task   |  Task   |
-/// +----------------+---------+---------+---------+
-/// | Contribution 1 |  Task   |  Task   |  Task   |
-/// +----------------+---------+---------+---------+
-/// | Contribution 0 |  Task   |  Task   |  Task   |
-/// +----------------+---------+---------+---------+
-/// | Chunk          | Chunk 0 | Chunk 1 | ...     |
-/// +----------------+---------+---------+---------+
-/// ```
-///
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub struct Task {
-    chunk_id: u64,
-    contribution_id: u64,
-}
-
-impl Task {
-    #[inline]
-    pub fn new(chunk_id: u64, contribution_id: u64) -> Self {
-        Self {
-            chunk_id,
-            contribution_id,
-        }
-    }
-
-    #[inline]
-    pub fn contains(&self, chunk_id: u64) -> bool {
-        self.chunk_id == chunk_id
-    }
-
-    #[inline]
-    pub fn chunk_id(&self) -> u64 {
-        self.chunk_id
-    }
-
-    #[inline]
-    pub fn contribution_id(&self) -> u64 {
-        self.contribution_id
-    }
-
-    #[inline]
-    pub fn to_tuple(&self) -> (u64, u64) {
-        (self.chunk_id, self.contribution_id)
-    }
-}
-
-impl Serialize for Task {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        serializer.serialize_str(&format!("{}/{}", self.chunk_id, self.contribution_id))
-    }
-}
-
-impl<'de> Deserialize<'de> for Task {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Task, D::Error> {
-        let s = String::deserialize(deserializer)?;
-
-        let mut task = s.split("/");
-        let chunk_id = task.next().ok_or(D::Error::custom("invalid chunk ID"))?;
-        let contribution_id = task.next().ok_or(D::Error::custom("invalid contribution ID"))?;
-        Ok(Task::new(
-            u64::from_str(&chunk_id).map_err(de::Error::custom)?,
-            u64::from_str(&contribution_id).map_err(de::Error::custom)?,
-        ))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::{coordinator_state::*, testing::prelude::*, CoordinatorState, SystemTimeSource};
-
-    use std::collections::HashSet;
-
-    #[test]
-    fn test_task() {
-        let task = Task::new(0, 1);
-        assert_eq!("\"0/1\"", serde_json::to_string(&task).unwrap());
-        assert_eq!(task, serde_json::from_str("\"0/1\"").unwrap());
-    }
-
-    #[test]
-    fn test_initialize_tasks() {
-        let bucket_id = 0;
-        let number_of_chunks = 3;
-        let number_of_contributors = 2;
-        let mut tasks = initialize_tasks(bucket_id, number_of_chunks, number_of_contributors).into_iter();
-        assert_eq!(Some(Task::new(0, 1)), tasks.next());
-        assert_eq!(Some(Task::new(1, 2)), tasks.next());
-        assert_eq!(Some(Task::new(2, 2)), tasks.next());
-
-        let bucket_id = 1;
-        let number_of_chunks = 3;
-        let number_of_contributors = 2;
-        let mut tasks = initialize_tasks(bucket_id, number_of_chunks, number_of_contributors).into_iter();
-        assert_eq!(Some(Task::new(1, 1)), tasks.next());
-        assert_eq!(Some(Task::new(2, 1)), tasks.next());
-        assert_eq!(Some(Task::new(0, 2)), tasks.next());
-    }
-
-    #[test]
-    fn test_initialize_tasks_unique() {
-        test_logger();
-
-        fn test_uniqueness_of_tasks(number_of_chunks: u64, number_of_contributors: u64) {
-            let mut all_tasks = HashSet::new();
-            for bucket_id in 0..number_of_contributors {
-                // trace!("Contributor {}", bucket_id);
-                let mut tasks = initialize_tasks(bucket_id, number_of_chunks, number_of_contributors).into_iter();
-                while let Some(task) = tasks.next() {
-                    assert!(all_tasks.insert(task));
-                }
-            }
-        }
-
-        for number_of_contributors in 1..32 {
-            trace!("{} contributors", number_of_contributors,);
-            for number_of_chunks in number_of_contributors..256 {
-                test_uniqueness_of_tasks(number_of_chunks, number_of_contributors);
-            }
-        }
-    }
 
     #[test]
     fn test_new() {
@@ -3819,9 +3639,9 @@ mod tests {
         for chunk_id in 0..contributor_lock_chunk_limit {
             // Fetch a pending task for the contributor.
             let task = state.fetch_task(&contributor, &time).unwrap();
-            assert_eq!((chunk_id as u64, 1), (task.chunk_id, task.contribution_id));
+            assert_eq!((chunk_id as u64, 1), (task.chunk_id(), task.contribution_id()));
 
-            state.acquired_lock(&contributor, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&contributor, task.chunk_id(), &time).unwrap();
             assert_eq!(0, state.pending_verification.len());
         }
 
@@ -3882,11 +3702,11 @@ mod tests {
             // Fetch a pending task for the contributor.
             let task = state.fetch_task(&contributor, &time).unwrap();
             let chunk_id = i as u64;
-            assert_eq!((chunk_id, 1), (task.chunk_id, task.contribution_id));
+            assert_eq!((chunk_id, 1), (task.chunk_id(), task.contribution_id()));
 
             state.acquired_lock(&contributor, chunk_id, &time).unwrap();
             state
-                .completed_task(&contributor, chunk_id, task.contribution_id, &time)
+                .completed_task(&contributor, chunk_id, task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(i + 1, state.pending_verification.len());
         }
@@ -3901,11 +3721,11 @@ mod tests {
         for i in 0..environment.verifier_lock_chunk_limit() {
             // Fetch a pending task for the verifier.
             let task = state.fetch_task(&verifier, &time).unwrap();
-            assert_eq!((i as u64, 1), (task.chunk_id, task.contribution_id));
+            assert_eq!((i as u64, 1), (task.chunk_id(), task.contribution_id()));
 
-            state.acquired_lock(&verifier, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&verifier, task.chunk_id(), &time).unwrap();
             state
-                .completed_task(&verifier, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&verifier, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(contributor_lock_chunk_limit - i - 1, state.pending_verification.len());
         }
@@ -3981,9 +3801,9 @@ mod tests {
             let expected_task1 = tasks1.next();
             assert_eq!(expected_task1, Some(&task));
 
-            state.acquired_lock(&contributor_1, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&contributor_1, task.chunk_id(), &time).unwrap();
             let assigned_verifier_1 = state
-                .completed_task(&contributor_1, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&contributor_1, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(1, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -3994,9 +3814,9 @@ mod tests {
             let expected_task2 = tasks2.next();
             assert_eq!(expected_task2, Some(&task));
 
-            state.acquired_lock(&contributor_2, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&contributor_2, task.chunk_id(), &time).unwrap();
             let assigned_verifier_2 = state
-                .completed_task(&contributor_2, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&contributor_2, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(2, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4006,9 +3826,9 @@ mod tests {
             let task = state.fetch_task(&verifier, &time).unwrap();
             assert_eq!(expected_task1, Some(&task));
 
-            state.acquired_lock(&verifier, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&verifier, task.chunk_id(), &time).unwrap();
             state
-                .completed_task(&verifier, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&verifier, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(1, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4017,9 +3837,9 @@ mod tests {
             let task = state.fetch_task(&verifier, &time).unwrap();
             assert_eq!(expected_task2, Some(&task));
 
-            state.acquired_lock(&verifier, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&verifier, task.chunk_id(), &time).unwrap();
             state
-                .completed_task(&verifier, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&verifier, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(0, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4110,9 +3930,9 @@ mod tests {
             let expected_task1 = tasks1.next();
             assert_eq!(expected_task1, Some(&task));
 
-            state.acquired_lock(&contributor_1, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&contributor_1, task.chunk_id(), &time).unwrap();
             let assigned_verifier = state
-                .completed_task(&contributor_1, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&contributor_1, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(1, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4121,9 +3941,9 @@ mod tests {
             let task = state.fetch_task(&assigned_verifier, &time).unwrap();
             assert_eq!(expected_task1, Some(&task));
 
-            state.acquired_lock(&assigned_verifier, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&assigned_verifier, task.chunk_id(), &time).unwrap();
             state
-                .completed_task(&assigned_verifier, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&assigned_verifier, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(0, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4164,9 +3984,9 @@ mod tests {
             let expected_task2 = tasks2.next();
             assert_eq!(expected_task2, Some(&task));
 
-            state.acquired_lock(&contributor_2, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&contributor_2, task.chunk_id(), &time).unwrap();
             let assigned_verifier = state
-                .completed_task(&contributor_2, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&contributor_2, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(1, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
@@ -4175,9 +3995,9 @@ mod tests {
             let task = state.fetch_task(&assigned_verifier, &time).unwrap();
             assert_eq!(expected_task2, Some(&task));
 
-            state.acquired_lock(&assigned_verifier, task.chunk_id, &time).unwrap();
+            state.acquired_lock(&assigned_verifier, task.chunk_id(), &time).unwrap();
             state
-                .completed_task(&assigned_verifier, task.chunk_id, task.contribution_id, &time)
+                .completed_task(&assigned_verifier, task.chunk_id(), task.contribution_id(), &time)
                 .unwrap();
             assert_eq!(0, state.pending_verification.len());
             assert!(!state.is_current_round_finished());
